The Answer!
Hello!
The answer to this puzzle is actually very simple, which is itself a testament to CMake :)

For the say library we add the add_library command to create a library named say
The library is supposed to be static, so we set the second parameter of add_library to STATIC
The third parameter is the list of the source files that make the say library

Along with creating a library, add_library creates an internal CMake target, which can be used
as a dependency for other targets in the CMake build.
Such as the hello executable!

For the hello executable we add the add_executable command to create an executable named hello
Similar to the add_library command we need to list the source files that make the hello executable,
also add_executable creates an internal CMake target,
but unlike the add_library command we don't need to specify what type of executable to create.

But! Creating the hello executable won't work on its own.
hello depends on the say libray.
To satisfy this dependency we add the target_link_libraries command, specifying the dependent
target, hello, in the first parameter and the target it depends on, say, in the second parameter

But still! This is not enough because although hello is linking the say library, it cannot find the say.h header!
We could add the include_directories command and specify ../lib/ as an include directoy,
but that relies on the say.h header remaining in that lib directory, which it may not.

Instead we are going to go back to the say library and add the target_include_directories command.
target_include_directories essentially tells all targets that depend on the say target where its headers are located.
The first parameter specifies the target to set include directories for.
The second parameter specifies that these include directories are PUBLIC and can be used to find headers for the library.
The third parameter is the location of the headers; we use the CMake variable CMAKE_CURRENT_SOURCE_DIR,
which basically means the directory that the say CMakeLists.txt file exists in.

Finally!
Running 'cmake .' in the root of this repo followed by 'make' should compile the lib/libsay.a library and the bin/hello executable

https://cmake.org/cmake/help/v3.2/command/add_library.html
https://cmake.org/cmake/help/v3.2/command/add_executable.html
https://cmake.org/cmake/help/v3.2/command/target_link_libraries.html
https://cmake.org/cmake/help/v3.2/command/include_directories.html
https://cmake.org/cmake/help/v3.2/command/target_include_directories.html